backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
deepface==0.0.86
tensorflow==2.11.0
numpy==1.24.3
python-dotenv==1.0.1
playwright==1.41.2
requests==2.31.0
